// project/src/types/hotel-type.ts
interface Location {
  latitude: number;
  longitude: number;
  zoom: number;
}

interface Host {
  avatarUrl: string;
  id: number;
  isPro: boolean;
  name: string;
}

interface City {
  location: Location;
  name: string;
}

export interface HotelOffer {
  bedrooms: number;
  city: City;
  description: string;
  goods: string[];
  host: Host;
  id: number;
  images: string[];
  isFavorite: boolean;
  isPremium: boolean;
  location: Location;
  maxAdults: number;
  previewImage: string;
  price: number;
  rating: number;
  title: string;
  type: string;
  }